requirements.txt
astroid==2.2.5
Babel==2.7.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.2
django-phonenumber-field==3.0.1
djangorestframework==3.9.4
djangorestframework-gis==0.14
drf-yasg==1.16.0
entrypoints==0.3
factory-boy==2.12.0
Faker==1.0.7
flake8==3.7.7
idna==2.8
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
phonenumbers==8.10.14
postgis==1.0.4
psycopg2==2.8.3
py-moneyed==0.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.3.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
ruamel.yaml==0.15.97
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
wrapt==1.11.2
